Refactor package structure
So each applet related class is under the `applet` package.
