+++ applet/bin/test/log4j.properties
@@ -1,0 +1,8 @@
+# Root Logger Option
+log4j.rootLogger=DEBUG, console
+
+# Redirect Log Messages To Console
+log4j.appender.console=org.apache.log4j.ConsoleAppender
+log4j.appender.console.Target=System.out
+log4j.appender.console.layout=org.apache.log4j.PatternLayout
+log4j.appender.console.layout.ConversionPattern=%-5p | %d{yyyy-MM-dd HH:mm:ss} | [%t] %C{2}:%L | %m%n
