+++ applet/src/main/java/applet/crypto/HmacSha256.java
@@ -1,4 +1,4 @@
-package crypto;
+package applet.crypto;
 
 import javacard.framework.ISOException;
 import javacard.framework.Util;
